+++ android/app/src/main/kotlin/com/eltatamanager/app/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.eltatamanager.app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()